menambahkan logika sistem dimana bisa menambahkan produk selain kepiting dan bisa menghitung kuantitas desimal

+++ app/src/main/java/com/example/crabsupply/data/model/Order.kt
@@ -8,7 +8,7 @@ data class Order(
     val productName: String = "",  // Barang yang dibeli
     val productSpecies: String = "", // Jenis (Bakau/Rajungan)
     val pricePerKg: Int = 0,       // Harga saat deal
-    val quantity: Int = 0,         // Jumlah Kg
+    val quantity: Double = 0.0,         // Jumlah Kg
     val totalPrice: Int = 0,       // Total Bayar (Harga x Qty)
     val status: String = "pending",// Status: pending, diproses, selesai, tolak
     val address: String = "",      // Alamat kirim
